# Back/swagger-spec.json
{"openapi":"3.0.0","paths":{"/users/me":{"get":{"operationId":"UsersController_getMe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UsersController_updateMe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_removeMe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"responses":{"200":{"description":""},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me/like/{idClothe}":{"put":{"operationId":"UsersController_likeClothe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"idClothe","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me/dislike/{idClothe}":{"put":{"operationId":"UsersController_dislikeClothe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"idClothe","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me/like":{"get":{"operationId":"UsersController_getlikeClothe","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/me/picture":{"put":{"operationId":"UsersController_updateMePicture","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":[],"properties":{"picture":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UsersController_findAll","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserWithoutPassword"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]},"post":{"operationId":"UsersController_createUserAdmin","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserAdminDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{uuid}":{"get":{"operationId":"UsersController_findOne","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UsersController_update","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{uuid}/picture":{"put":{"operationId":"UsersController_updatePicture","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["picture"],"properties":{"picture":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithoutPassword"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Users"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthReturnDto"}}}}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthReturnDto"}}}}},"tags":["Auth"]}},"/files/download/{id}":{"get":{"operationId":"FilesController_download","summary":"","description":"Download file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string","format":"binary"}}}}},"tags":["Files"]}},"/files/show/{id}":{"get":{"operationId":"FilesController_show","summary":"","description":"Show file","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string","format":"binary"}}}}},"tags":["Files"]}},"/files":{"post":{"operationId":"FilesController_create","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFileDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Files"],"security":[{"bearer":[]}]},"get":{"operationId":"FilesController_findAll","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Files"],"security":[{"bearer":[]}]}},"/files/{id}":{"get":{"operationId":"FilesController_findOne","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Files"],"security":[{"bearer":[]}]},"delete":{"operationId":"FilesController_remove","summary":"","description":"## Access\n### To access this route need to be authentified and have one of this role [Admin, SuperAdmin].","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Files"],"security":[{"bearer":[]}]}},"/clothes":{"post":{"operationId":"ClothesController_create","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClotheDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]},"get":{"operationId":"ClothesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}}}}}},"tags":["Clothes"]}},"/clothes/{id}":{"get":{"operationId":"ClothesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}}},"tags":["Clothes"]},"put":{"operationId":"ClothesController_update","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClotheDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]},"delete":{"operationId":"ClothesController_remove","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]}},"/clothes/{id}/liked":{"get":{"operationId":"ClothesController_findOneLiked","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]}},"/clothes/{id}/order":{"put":{"operationId":"ClothesController_order","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]}},"/clothes/{id}/picture":{"put":{"operationId":"ClothesController_updatePicture","summary":"","description":"## Access\n### To access this route need to be authentified.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["picture"],"properties":{"picture":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClotheEntity"}}}},"401":{"description":"You need to be logged to access this route."},"403":{"description":"You don't have the right to access this route."}},"tags":["Clothes"],"security":[{"bearer":[]}]}}},"info":{"title":"Pwa epitech backend API","description":"The test API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"FileEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"path":{"type":"string"},"mimetype":{"type":"string"},"downloadLink":{"type":"string"},"showLink":{"type":"string"}},"required":["id","name","path","mimetype","downloadLink","showLink"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"role":{"enum":["SuperAdmin","Admin","User"],"type":"string"},"picture":{"$ref":"#/components/schemas/FileEntity"},"clotheLike":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}},"swSub":{"type":"string"}},"required":["id","username","password","email","role"]},"ClotheEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"clotheAvaible":{"type":"number"},"picture":{"$ref":"#/components/schemas/FileEntity"},"user":{"$ref":"#/components/schemas/UserEntity"},"longitude":{"type":"number"},"latitude":{"type":"number"},"liked":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}},"required":["id","name","clotheAvaible","user"]},"UserWithoutPassword":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"role":{"enum":["SuperAdmin","Admin","User"],"type":"string"},"picture":{"$ref":"#/components/schemas/FileEntity"},"clotheLike":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}},"swSub":{"type":"string"}},"required":["id","username","email","role"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string"},"clotheLike":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}},"swSub":{"type":"string"}}},"UserClotheEntity":{"type":"object","properties":{"userId":{"type":"string"},"clotheId":{"type":"string"},"clothes":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}},"users":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}},"required":["userId","clotheId","clothes","users"]},"CreateUserAdminDto":{"type":"object","properties":{"role":{"enum":["SuperAdmin","Admin","User"],"type":"string"},"username":{"type":"string","minLength":3,"maxLength":30},"password":{"type":"string","minLength":3,"maxLength":30},"email":{"type":"string"}},"required":["role","username","password","email"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","minLength":3,"maxLength":30},"password":{"type":"string","minLength":3,"maxLength":30},"email":{"type":"string"}},"required":["username","password","email"]},"AuthReturnDto":{"type":"object","properties":{"token":{"type":"string"},"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"role":{"enum":["SuperAdmin","Admin","User"],"type":"string"},"picture":{"$ref":"#/components/schemas/FileEntity"},"clotheLike":{"type":"array","items":{"$ref":"#/components/schemas/ClotheEntity"}},"swSub":{"type":"string"}},"required":["token","id","username","email","role"]},"LoginDto":{"type":"object","properties":{"password":{"type":"string","minLength":3,"maxLength":30},"username":{"type":"string"}},"required":["password","username"]},"CreateFileDto":{"type":"object","properties":{"name":{"type":"string"},"path":{"type":"string"},"mimetype":{"type":"string"}},"required":["name","path","mimetype"]},"CreateClotheDto":{"type":"object","properties":{"name":{"type":"string"},"clotheAvaible":{"type":"number","minimum":0},"longitude":{"type":"number"},"latitude":{"type":"number"}},"required":["name","clotheAvaible"]},"UpdateClotheDto":{"type":"object","properties":{"name":{"type":"string"},"clotheAvaible":{"type":"number","minimum":0},"longitude":{"type":"number"},"latitude":{"type":"number"}}}}}}